+++ Tests/Fixtures/UnusedParameterFixtures/testNestedVariable.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+class SyntaxFixture24 {
+    func myFunc(param: String) {
+        var innerProperty: String { param }
+    }
+}